feat: implement Room DB + Repository for 12th week

+++ app/src/main/java/com/smwu/bigsister/data/repository/ReservationRepository.kt
@@ -27,7 +27,7 @@ class ReservationRepository @Inject constructor(
         reservationDao.insertReservation(reservation)
     }
 
-    suspend fun deleteReservation(id: Long) {
+    suspend fun deleteReservation(id: Int) {
         reservationDao.deleteReservationById(id)
     }
 }
